refactor(FR-3673): make the status buckets a plain helper instead of a hook
Being used in three places does not by itself justify a custom hook, and no
existing hook in the tree exists solely to wrap a `supports()` gate — the
ones that call it (useCSVExport, useStartSession, useWebUIMenuItems) each
have their own substantive job and read the flag along the way.

`getSessionV2StatusBuckets(supportsPreemptionStatuses)` is pure, so each
consumer reads the client itself and passes the boolean in — the shape
`buildModelDefinitionInput` already uses in AdminDeploymentPresetSettingPage.
The buckets still live in one module, so the duplication this issue asked to
remove stays removed. The test drops its hook mock and calls the function
directly.
